Implement /detailed_diagnostic for LSP servers
Seems like the original LSP implementation forgot about
/detailed_diagnostic. Later, clangd completer implemented locally what
works for all LSP completers.

This commit moves the code from ClangdCompleter to
LanguageServerCompleter and adds a few tests.
