+++ ycmd/tests/clangd/utilities_test.py
@@ -34,37 +34,6 @@ from ycmd.tests.test_utils import BuildRequest
 from ycmd.user_options_store import DefaultOptions
 
 
-def _TupleToLSPRange( tuple ):
-  return { 'line': tuple[ 0 ], 'character': tuple[ 1 ] }
-
-
-def _Check_Distance( point, start, end, expected ):
-  point = _TupleToLSPRange( point )
-  start = _TupleToLSPRange( start )
-  end = _TupleToLSPRange( end )
-  range = { 'start': start, 'end': end }
-  result = clangd_completer.DistanceOfPointToRange( point, range )
-  eq_( result, expected )
-
-
-def ClangdCompleter_DistanceOfPointToRange_SingleLineRange_test():
-  # Point to the left of range.
-  _Check_Distance( ( 0, 0 ), ( 0, 2 ), ( 0, 5 ) , 2 )
-  # Point inside range.
-  _Check_Distance( ( 0, 4 ), ( 0, 2 ), ( 0, 5 ) , 0 )
-  # Point to the right of range.
-  _Check_Distance( ( 0, 8 ), ( 0, 2 ), ( 0, 5 ) , 3 )
-
-
-def ClangdCompleter_DistanceOfPointToRange_MultiLineRange_test():
-  # Point to the left of range.
-  _Check_Distance( ( 0, 0 ), ( 0, 2 ), ( 3, 5 ) , 2 )
-  # Point inside range.
-  _Check_Distance( ( 1, 4 ), ( 0, 2 ), ( 3, 5 ) , 0 )
-  # Point to the right of range.
-  _Check_Distance( ( 3, 8 ), ( 0, 2 ), ( 3, 5 ) , 3 )
-
-
 def ClangdCompleter_GetClangdCommand_NoCustomBinary_test():
   user_options = DefaultOptions()
